# Essential Elements of an Effective PR Description Checklist

- [ ] The purpose of the PR, such as "Fix some issue (link existing issues this PR will resolve)".
- [ ] The test plan, such as providing test command.
- [ ] The test results, such as pasting the results comparison before and after, or e2e results


## Propose 
This PR enables the ViT backend selection in the Qwen2 and Qwen2.2 vision language models to support the `FLASH_ATTN` backend on the ROCm platform. It uses the flash_attn_varlen_func from vLLM and additionally provides support for the AITER flash attention implementation (flash_attn_varlen_func from AITER).

## Test Plan and Results

Use [mistral-evals](https://github.com/mistralai/mistral-evals/tree/main) repo for accuracy evaluation.

eval result on `Qwen/Qwen2.5-VL-7B-Instruct` model

Result of using `vllm flash attention` for ViT backend
{
    "explicit_prompt_relaxed_correctness": 0.8632,
    "anywhere_in_answer_relaxed_correctness": 0.8632
}

Result of using `aiter flash attention` for ViT backend
{
    "explicit_prompt_relaxed_correctness": 0.8644,
    "anywhere_in_answer_relaxed_correctness": 0.8648
}

## Benchmark

Benchmark results on `Qwen/Qwen2.5-VL-7B-Instruct`

| Metric                              | Base           | Flash Attention | Aiter Flash Attention (updated) |
|--------------------------------------|:--------------:|:---------------:|:-------------------------------:|
| **Successful requests**              | 1000           | 1000            | 1000                            |
| **Benchmark duration (s)**           | 669.11         | 438.96          | 439.20                          |
| **Total input tokens**               | 94,327         | 94,327          | 94,327                          |
| **Total generated tokens**           | 114,147        | 114,130         | 113,498                         |
| **Request throughput (req/s)**       | 1.49           | 2.28            | 2.28                            |
| **Output token throughput (tok/s)**  | 170.60         | 260.00          | 258.42                          |
| **Total token throughput (tok/s)**   | 311.57         | 474.89          | 473.19                          |
| **Mean TTFT (ms)**                   | 304,688.19     | 195,171.21      | 192,163.31                      |
| **Median TTFT (ms)**                 | 285,449.09     | 169,784.12      | 167,927.25                      |
| **P99 TTFT (ms)**                    | 653,794.14     | 426,574.61      | 432,378.60                      |
| **Mean TPOT (ms)**                   | 3,273.63       | 2,178.81        | 2,201.65                        |
| **Median TPOT (ms)**                 | 3,336.22       | 2,287.00        | 2,288.36                        |
| **P99 TPOT (ms)**                    | 7,618.51       | 4,620.32        | 4,606.39                        |
| **Mean ITL (ms)**                    | 2,996.72       | 1,999.78        | 2,029.19                        |
| **Median ITL (ms)**                  | 72.78          | 64.97           | 62.44                           |
| **P99 ITL (ms)**                     | 8,251.64       | 8,252.79        | 8,166.49                        |


**BEFORE SUBMITTING, PLEASE READ <https://docs.vllm.ai/en/latest/contributing>** (anything written below this line will be removed by GitHub Actions)